Striver Cp Sheet

+++ Codeforces/1107B - Digital root.cpp
@@ -1,7 +1,7 @@
 #include <bits/stdc++.h>
 using namespace std;
 void solve(){
-    int k, x;
+    long long k, x;
     cin>>k>>x;
     cout<<(k-1) * 9 + x<<endl;
 }
@@ -15,3 +15,4 @@ int main(){
 
     return 0;
 }
+
